Split off tests of variable methods to new class
